+++ firmware/controllers/engine_cycle/spark_logic.cpp
@@ -9,11 +9,6 @@
 
 #include "spark_logic.h"
 
-// dependency injection
-#include "engine_state.h"
-#include "rpm_calculator_api.h"
-// end of injection
-
 #include "utlist.h"
 #include "event_queue.h"
 
@@ -85,8 +80,7 @@ static void prepareCylinderIgnitionSchedule(angle_t dwellAngleDuration, floatms_
 									// Pull any extra timing for knock retard
 									- engine->module<KnockController>()->getKnockRetard();
 
-	// TODO: Log "true" per-cylinder timing here #76
-	//engine->outputChannels......[event->cylinderNumber] = finalIgnitionTiming;
+	engine->outputChannels.ignitionAdvanceCyl[event->cylinderNumber] = finalIgnitionTiming;
 
 	angle_t sparkAngle =
 		// Negate because timing *before* TDC, and we schedule *after* TDC